Rename data module to tables, update package for constants, update naming convention for ColumnFamilyConstants
