+++ core/pom.xml
@@ -21,7 +21,7 @@
         <module>query</module>
         <module>utils</module>
         <module>base-rest-responses</module>
-        <module>data</module>
+        <module>tables</module>
     </modules>
     <dependencies>
         <dependency>
