+++ warehouse/ingest-core/src/main/java/datawave/ingest/table/config/DateIndexTableConfigHelper.java
@@ -15,8 +15,8 @@ import org.apache.hadoop.io.Text;
 import org.apache.hadoop.util.StringUtils;
 import org.apache.log4j.Logger;
 
-import datawave.constants.ColumnFamilyConstants;
-import datawave.constants.LocalityGroupConstants;
+import datawave.table.constants.ColumnFamilyConstants;
+import datawave.table.constants.LocalityGroupConstants;
 
 public class DateIndexTableConfigHelper extends AbstractTableConfigHelper {
 
@@ -39,8 +39,8 @@ public class DateIndexTableConfigHelper extends AbstractTableConfigHelper {
         }
 
         String localityGroupsConf = null;
-        localityGroupsConf = conf.get(LOCALITY_GROUPS, LocalityGroupConstants.FULL_CONTENT_LOCALITY + ':' + ColumnFamilyConstants.FULL_CONTENT_NAME + ','
-                        + LocalityGroupConstants.TERM_FREQUENCY_LOCALITY + ':' + ColumnFamilyConstants.TERM_FREQUENCY_NAME);
+        localityGroupsConf = conf.get(LOCALITY_GROUPS, LocalityGroupConstants.FULL_CONTENT_LOCALITY + ':' + ColumnFamilyConstants.FULL_CONTENT + ','
+                        + LocalityGroupConstants.TERM_FREQUENCY_LOCALITY + ':' + ColumnFamilyConstants.TERM_FREQUENCY);
         for (String localityGroupDefConf : StringUtils.split(localityGroupsConf)) {
             String[] localityGroupDef = StringUtils.split(localityGroupDefConf, '\\', ':');
             Set<Text> families = localityGroups.get(localityGroupDef[0]);
