+++ warehouse/query-core/src/main/java/datawave/query/tables/content/ContentQueryLogic.java
@@ -22,7 +22,6 @@ import org.apache.log4j.Logger;
 
 import com.google.common.collect.Lists;
 
-import datawave.constants.ColumnFamilyConstants;
 import datawave.core.common.connection.AccumuloConnectionFactory;
 import datawave.core.query.configuration.GenericQueryConfiguration;
 import datawave.core.query.logic.BaseQueryLogic;
@@ -37,6 +36,7 @@ import datawave.query.QueryParameters;
 import datawave.query.config.ContentQueryConfiguration;
 import datawave.query.tables.ScannerFactory;
 import datawave.query.transformer.ContentQueryTransformer;
+import datawave.table.constants.ColumnFamilyConstants;
 import datawave.webservice.query.exception.QueryException;
 
 /**
@@ -228,7 +228,7 @@ public class ContentQueryLogic extends BaseQueryLogic<Entry<Key,Value>> implemen
                     log.debug("Received pieces: " + shardId + ", " + datatype + ", " + uid);
 
                     // Create and add a Range
-                    final String cf = ColumnFamilyConstants.FULL_CONTENT_NAME;
+                    final String cf = ColumnFamilyConstants.FULL_CONTENT;
                     final String cq = datatype + Constants.NULL_BYTE_STRING + uid;
                     final Key startKey = new Key(shardId, cf, cq + Constants.NULL_BYTE_STRING);
                     final Key endKey = new Key(shardId, cf, cq + endKeyTerminator);
